Move OCP on AWS to ocp_aws provider folder + add in new fields
