+++ koku/reporting/models.py
@@ -28,8 +28,7 @@ from reporting.provider.aws.models import (AWSAccountAlias,                    #
                                            AWSCostEntryPricing,                # noqa: F401
                                            AWSCostEntryProduct,                # noqa: F401
                                            AWSCostEntryReservation)            # noqa: F401
-from reporting.provider.ocp.models import (OCPAWSCostLineItemDailySummary,     # noqa: F401
-                                           OCPStorageLineItem,                 # noqa: F401
+from reporting.provider.ocp.models import (OCPStorageLineItem,                 # noqa: F401
                                            OCPStorageLineItemDaily,            # noqa: F401
                                            OCPStorageLineItemDailySummary,     # noqa: F401
                                            OCPUsageLineItem,                   # noqa: F401
@@ -39,3 +38,6 @@ from reporting.provider.ocp.models import (OCPAWSCostLineItemDailySummary,     #
                                            OCPUsagePodLabelSummary,            # noqa: F401
                                            OCPUsageReport,                     # noqa: F401
                                            OCPUsageReportPeriod)               # noqa: F401
+from reporting.provider.ocp_aws.models import (OCPAWSCostLineItemDailySummary, # noqa: F401
+                                               OCPAWSStorageLineItemDaily,     # noqa: F401
+                                               OCPAWSUsageLineItemDaily)       # noqa: F401
